+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [Unreleased]
 [Unreleased]: https://github.com/JakeWharton/mosaic/compare/0.16.0...HEAD
 
-Nothing yet!
+Changed:
+- Unsolicited focus events are now ignored unless the terminal has reported that it supports focus.
+  This should have no real impact on anything, except that now the `Terminal.capabilities.focus` value can now be trusted to indicate whether `Terminal.state.focus` will ever change.
 
 
 ## [0.17.0] - 2025-04-11